Refactor views into separate directories, update related files and improved documentation
